+ 1. Цвета - регистрация авторизация, изменения профиля + 2. Проблема с датами (в каталоге маршрутов, сохранение экземпляров всех и не всех) + 3. редирект у профиля если не авторизован + 4. убрать кнопку навверх в редакторе точек + 5. изменение адреса сделать нередактируемым + 6. вывод ошибки при сохранении точек (можно сделать валидацию названия) + 7. не показывает изображения у ключевых точек + 8. нормально не сохраняется если нет изображений у ключевых точек + 9. не удаляется точка если она единственная + 10. маршрут не показывается в общей информации о маршруте + 11. Красным почеркивать экз. маршрута при переходе из таблицы на dashboard

+++ src/Core/Services/RouteExampleService.cs
@@ -54,8 +54,19 @@ public class RouteExampleService(
         return RouteExampleDto.FromEntity(result);
     }
 
-    public async Task DeleteAsync(long id) =>
+    public async Task DeleteAsync(long id)
+    {
+        var routeExample = await routeExampleRepository.Items.Include(x => x.RouteExampleRecords).FirstOrDefaultAsync(x => x.Id == id);
+        if (routeExample is null)
+        {
+            throw new ApiEntityNotFoundException("Не удалось найти экземпляр маршрута");
+        }
+        else if (routeExample.RouteExampleRecords.Any())
+        {
+            throw new ApiValidationException("Невозможно удалить экземпляр маршрута, на котором есть записи");
+        }
         await routeExampleRepository.DeleteAsync(id);
+    }
 
     public async Task<RouteExampleDto> GetAsync(long id)
     {
@@ -76,9 +87,21 @@ public class RouteExampleService(
 
     public async Task<IEnumerable<RouteExampleDto>> GetExamplesByRouteId(long routeId)
     {
-        return await routeExampleRepository.Items
+        var recordsQuery = routeExampleRepository.Items
             .Where(x => x.RouteId == routeId)
-            .Select(x => RouteExampleDto.FromEntity(x)).ToListAsync();
+            .Include(x => x.RouteExampleRecords);
+
+        var records = await recordsQuery.ToListAsync();
+
+        var result = records.Adapt<List<RouteExampleDto>>();
+        var lengths = records.Select(x => x.RouteExampleRecords.Count).ToList();
+
+        for (var i = 0; i < records.Count; i++)
+        {
+            result[i].CountRecords = lengths[i];
+        }
+
+        return result;
     }
 
     public async Task<RouteExampleDto> CreateOrUpdateAsync(RouteExampleCreateOrUpdateRequest request)
